Revert "Hunger and thirst HUDs (#18066)"
This reverts commit 67e4b626c1c490f98950add60f42c4465cf049a1.
