+++ Content.Server/Nutrition/EntitySystems/ThirstSystem.cs
@@ -1,71 +1,33 @@
-using Content.Shared.Alert;
-using Content.Shared.Movement.Components;
-using Content.Shared.Movement.Systems;
-using Content.Shared.Nutrition.Components;
-using Content.Shared.Rejuvenate;
+using Content.Server.Nutrition.Components;
 using JetBrains.Annotations;
-using Robust.Shared.GameStates;
 using Robust.Shared.Random;
-using Robust.Shared.Timing;
+using Content.Shared.Movement.Components;
+using Content.Shared.Alert;
+using Content.Shared.Movement.Systems;
+using Content.Shared.Rejuvenate;
 
-namespace Content.Shared.Nutrition.EntitySystems
+namespace Content.Server.Nutrition.EntitySystems
 {
+    [UsedImplicitly]
     public sealed class ThirstSystem : EntitySystem
     {
-        [Dependency] private readonly IGameTiming _timing = default!;
         [Dependency] private readonly IRobustRandom _random = default!;
         [Dependency] private readonly AlertsSystem _alerts = default!;
         [Dependency] private readonly MovementSpeedModifierSystem _movement = default!;
         [Dependency] private readonly SharedJetpackSystem _jetpack = default!;
 
         private ISawmill _sawmill = default!;
+        private float _accumulatedFrameTime;
 
         public override void Initialize()
         {
             base.Initialize();
 
             _sawmill = Logger.GetSawmill("thirst");
-            SubscribeLocalEvent<ThirstComponent, ComponentGetState>(OnGetState);
-            SubscribeLocalEvent<ThirstComponent, ComponentHandleState>(OnHandleState);
-            SubscribeLocalEvent<ThirstComponent, EntityUnpausedEvent>(OnUnpaused);
             SubscribeLocalEvent<ThirstComponent, RefreshMovementSpeedModifiersEvent>(OnRefreshMovespeed);
-            SubscribeLocalEvent<ThirstComponent, ComponentShutdown>(OnShutdown);
             SubscribeLocalEvent<ThirstComponent, ComponentStartup>(OnComponentStartup);
             SubscribeLocalEvent<ThirstComponent, RejuvenateEvent>(OnRejuvenate);
         }
-
-        private void OnGetState(EntityUid uid, ThirstComponent component, ref ComponentGetState args)
-        {
-            args.State = new ThirstComponentState(component.BaseDecayRate,
-                component.ActualDecayRate,
-                component.CurrentThirstThreshold,
-                component.LastThirstThreshold,
-                component.CurrentThirst,
-                component.NextUpdateTime);
-        }
-
-        private void OnHandleState(EntityUid uid, ThirstComponent component, ref ComponentHandleState args)
-        {
-            if (args.Current is not ThirstComponentState state)
-                return;
-            component.BaseDecayRate = state.BaseDecayRate;
-            component.ActualDecayRate = state.ActualDecayRate;
-            component.CurrentThirstThreshold = state.CurrentThirstThreshold;
-            component.LastThirstThreshold = state.LastThirstThreshold;
-            component.CurrentThirst = state.CurrentThirst;
-            component.NextUpdateTime = state.NextUpdateTime;
-        }
-
-        private void OnUnpaused(EntityUid uid, ThirstComponent component, ref EntityUnpausedEvent args)
-        {
-            component.NextUpdateTime += args.PausedTime;
-        }
-
-        private void OnShutdown(EntityUid uid, ThirstComponent component, ComponentShutdown args)
-        {
-            _alerts.ClearAlertCategory(uid, AlertCategory.Thirst);
-        }
-
         private void OnComponentStartup(EntityUid uid, ThirstComponent component, ComponentStartup args)
         {
             // Do not change behavior unless starting value is explicitly defined
@@ -187,26 +149,23 @@ namespace Content.Shared.Nutrition.EntitySystems
                     throw new ArgumentOutOfRangeException($"No thirst threshold found for {component.CurrentThirstThreshold}");
             }
         }
-
         public override void Update(float frameTime)
         {
-            base.Update(frameTime);
+            _accumulatedFrameTime += frameTime;
 
-            var query = EntityQueryEnumerator<ThirstComponent>();
-            while (query.MoveNext(out var thirst))
+            if (_accumulatedFrameTime > 1)
             {
-                if (_timing.CurTime < thirst.NextUpdateTime)
-                    continue;
-                thirst.NextUpdateTime = _timing.CurTime + thirst.UpdateRate;
-
-                UpdateThirst(thirst, -thirst.ActualDecayRate);
-                var calculatedThirstThreshold = GetThirstThreshold(thirst, thirst.CurrentThirst);
-                if (calculatedThirstThreshold != thirst.CurrentThirstThreshold)
+                foreach (var component in EntityManager.EntityQuery<ThirstComponent>())
                 {
-                    thirst.CurrentThirstThreshold = calculatedThirstThreshold;
-                    UpdateEffects(thirst);
+                    UpdateThirst(component, - component.ActualDecayRate);
+                    var calculatedThirstThreshold = GetThirstThreshold(component, component.CurrentThirst);
+                    if (calculatedThirstThreshold != component.CurrentThirstThreshold)
+                    {
+                        component.CurrentThirstThreshold = calculatedThirstThreshold;
+                        UpdateEffects(component);
+                    }
                 }
-                Dirty(thirst);
+                _accumulatedFrameTime -= 1;
             }
         }
     }
