--- Content.Shared/Overlays/ShowHungerIconsComponent.cs
@@ -1,8 +1,0 @@
-﻿namespace Content.Shared.Overlays
-{
-    /// <summary>
-    /// This component allows you to see the hungriness of mobs.
-    /// </summary>
-    [RegisterComponent]
-    public sealed class ShowHungerIconsComponent : Component { }
-}
