--- Content.Shared/Overlays/ShowSecurityIconsComponent.cs
@@ -1,8 +1,0 @@
-namespace Content.Shared.Overlays
-{
-    /// <summary>
-    /// This component allows you to see job icons above mobs.
-    /// </summary>
-    [RegisterComponent]
-    public sealed class ShowSecurityIconsComponent : Component { }
-}
