--- Content.Shared/Overlays/ShowThirstIconsComponent.cs
@@ -1,8 +1,0 @@
-﻿namespace Content.Shared.Overlays
-{
-    /// <summary>
-    /// This component allows you to see the thirstiness of mobs.
-    /// </summary>
-    [RegisterComponent]
-    public sealed class ShowThirstIconsComponent : Component { }
-}
