+++ Resources/Prototypes/Entities/Clothing/Eyes/glasses.yml
@@ -1,5 +1,16 @@
 - type: entity
   parent: ClothingEyesBase
+  id: ClothingEyesGlassesBeer
+  name: beer goggles
+  description: A pair of sunglasses outfitted with apparatus to scan reagents, as well as providing an innate understanding of liquid viscosity while in motion.
+  components:
+  - type: Sprite
+    sprite: Clothing/Eyes/Glasses/beergoggles.rsi
+  - type: Clothing
+    sprite: Clothing/Eyes/Glasses/beergoggles.rsi
+
+- type: entity
+  parent: ClothingEyesBase
   id: ClothingEyesGlassesGar
   name: gar glasses
   description: Go beyond impossible and kick reason to the curb!
