+++ Resources/Prototypes/Roles/Jobs/Civilian/bartender.yml
@@ -22,7 +22,6 @@
   id: BartenderGear
   equipment:
     head: ClothingHeadHatTophat
-    eyes: ClothingEyesGlassesBeer
     jumpsuit: ClothingUniformJumpsuitBartender
     outerClothing: ClothingOuterVest
     back: ClothingBackpackFilled
