--- Resources/Prototypes/StatusEffects/hunger.yml
@@ -1,49 +1,0 @@
-#Hunger
-- type: statusIcon
-  id: HungerIcon_Overfed
-  priority: 5
-  icon:
-    sprite: Interface/Misc/food_icons.rsi
-    state: overfed
-  locationPreference: Right
-
-- type: statusIcon
-  id: HungerIcon_Peckish
-  priority: 5
-  icon:
-    sprite: Interface/Misc/food_icons.rsi
-    state: peckish
-  locationPreference: Right
-
-- type: statusIcon
-  id: HungerIcon_Starving
-  priority: 5
-  icon:
-    sprite: Interface/Misc/food_icons.rsi
-    state: starving
-  locationPreference: Right
-
-#Thirst
-- type: statusIcon
-  id: ThirstIcon_Overhydrated
-  priority: 5
-  icon:
-    sprite: Interface/Misc/food_icons.rsi
-    state: overhydrated
-  locationPreference: Left
-
-- type: statusIcon
-  id: ThirstIcon_Thirsty
-  priority: 5
-  icon:
-    sprite: Interface/Misc/food_icons.rsi
-    state: thirsty
-  locationPreference: Left
-
-- type: statusIcon
-  id: ThirstIcon_Parched
-  priority: 5
-  icon:
-    sprite: Interface/Misc/food_icons.rsi
-    state: parched
-  locationPreference: Left
