--- Resources/Prototypes/StatusEffects/job.yml
@@ -1,375 +1,0 @@
-- type: statusIcon
-  id: JobIcon_Detective
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Detective
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_QuarterMaster
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: QuarterMaster
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Botanist
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Botanist
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Boxer
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Boxer
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_AtmosphericTechnician
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: AtmosphericTechnician
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Nanotrasen
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Nanotrasen
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Prisoner
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Prisoner
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Janitor
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Janitor
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Chemist
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Chemist
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_StationEngineer
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: StationEngineer
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_SecurityOfficer
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: SecurityOfficer
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_NoId
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: NoId
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_ChiefMedicalOfficer
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: ChiefMedicalOfficer
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Roboticist
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Roboticist
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Chaplain
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Chaplain
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Lawyer
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Lawyer
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Unknown
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Unknown
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Librarian
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Librarian
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_CargoTechnician
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: CargoTechnician
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Scientist
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Scientist
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_ResearchAssistant
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: ResearchAssistant
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Geneticist
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Geneticist
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Clown
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Clown
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Captain
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Captain
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_HeadOfPersonnel
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: HeadOfPersonnel
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Virologist
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Virologist
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_ShaftMiner
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: ShaftMiner
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Passenger
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Passenger
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_ChiefEngineer
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: ChiefEngineer
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Bartender
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Bartender
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_HeadOfSecurity
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: HeadOfSecurity
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Brigmedic
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Brigmedic
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_MedicalDoctor
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: MedicalDoctor
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Paramedic
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Paramedic
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Chef
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Chef
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Warden
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Warden
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_ResearchDirector
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: ResearchDirector
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Mime
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Mime
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Musician
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Musician
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Reporter
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Reporter
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Psychologist
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Psychologist
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_MedicalIntern
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: MedicalIntern
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_TechnicalAssistant
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: TechnicalAssistant
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_ServiceWorker
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: ServiceWorker
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_SecurityCadet
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: SecurityCadet
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Zombie
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Zombie
-  locationPreference: Left
-
-- type: statusIcon
-  id: JobIcon_Zookeeper
-  priority: 1
-  icon:
-    sprite: Interface/Misc/job_icons.rsi
-    state: Zookeeper
-  locationPreference: Left
