+++ Resources/Textures/Clothing/Eyes/Glasses/beergoggles.rsi/equipped-EYES.png
