+++ Resources/Textures/Clothing/Eyes/Glasses/beergoggles.rsi/icon.png
