+++ Resources/Textures/Clothing/Eyes/Glasses/beergoggles.rsi/meta.json
