+++ Resources/Textures/Clothing/Eyes/Glasses/friedoniongoggles.rsi/equipped-EYES.png
