+++ Resources/Textures/Clothing/Eyes/Glasses/friedoniongoggles.rsi/icon.png
