+++ Resources/Textures/Clothing/Eyes/Glasses/friedoniongoggles.rsi/meta.json
